Replace SyncStatus with enhanced UpdateStatus everywhere.
