+++ pixeltable/catalog/update_status.py
@@ -65,6 +65,9 @@ class UpdateStatus:
     # stats for changes cascaded to other tables
     cascade_row_count_stats: RowCountStats = field(default_factory=RowCountStats)
 
+    # stats for the rows affected by the operation in an external store
+    ext_row_count_stats: RowCountStats = field(default_factory=RowCountStats)
+
     @property
     def num_rows(self) -> int:
         return self.row_count_stats.num_rows + self.cascade_row_count_stats.num_rows
@@ -87,6 +90,7 @@ class UpdateStatus:
             cols_with_excs=self.cols_with_excs,
             row_count_stats=self.row_count_stats.insert_to_update(),
             cascade_row_count_stats=self.cascade_row_count_stats.insert_to_update(),
+            ext_row_count_stats=self.ext_row_count_stats,
         )
 
     def to_cascade(self) -> 'UpdateStatus':
@@ -99,6 +103,7 @@ class UpdateStatus:
             cols_with_excs=self.cols_with_excs,
             row_count_stats=RowCountStats(),
             cascade_row_count_stats=self.cascade_row_count_stats + self.row_count_stats,
+            ext_row_count_stats=self.ext_row_count_stats,
         )
 
     def __add__(self, other: 'UpdateStatus') -> UpdateStatus:
@@ -110,6 +115,7 @@ class UpdateStatus:
             cols_with_excs=list(dict.fromkeys(self.cols_with_excs + other.cols_with_excs)),
             row_count_stats=self.row_count_stats + other.row_count_stats,
             cascade_row_count_stats=self.cascade_row_count_stats + other.cascade_row_count_stats,
+            ext_row_count_stats=self.ext_row_count_stats + other.ext_row_count_stats,
         )
 
     @property
@@ -148,3 +154,26 @@ class UpdateStatus:
         if stats.num_excs > 0:
             messages.append(self.__cnt_str(stats.num_excs, 'exception'))
         p.text(', '.join(messages) + '.' if len(messages) > 0 else 'No rows affected.')
+
+    @property
+    def pxt_rows_updated(self) -> int:
+        """
+        Returns the number of Pixeltable rows that were updated as a result of the operation.
+        """
+        return (self.row_count_stats + self.cascade_row_count_stats).upd_rows
+
+    @property
+    def external_rows_updated(self) -> int:
+        return self.ext_row_count_stats.upd_rows
+
+    @property
+    def external_rows_created(self) -> int:
+        return self.ext_row_count_stats.ins_rows
+
+    @property
+    def external_rows_deleted(self) -> int:
+        return self.ext_row_count_stats.del_rows
+
+    @property
+    def ext_num_rows(self) -> int:
+        return self.ext_row_count_stats.num_rows
